fix: add initbg, otherwise vectortilesurl is set after setting bg confs
